[TASK] Dissolve NodeConverter
NodeConverter is bound to TemplateCompiler: When
templates are compiled, NodeConverter is called
with each node to turn it into PHP code.

With Fluid prior to v2, NodeConverter has been
inlined in TemplateCompiler, and has been
extracted with v2 to an own class.

This was not a great decision: NodeConverter is
essentially a switch-case structure for single
nodes. Instead of abstracting this to an own class,
single nodes should "know" how they are compiled.

The patch does exactly this: Move all sub methods
of NodeConverter to their distinct nodes and drop
NodeConverter entirely.

NodeInterface is thus extended with a convert()
method. Strictly speaking, the patch is breaking
due to this NodeInterface change. In practice,
all known consumers extend AbstractNode, which
comes with a default implementation that keeps
b/w compatibility.

Also, no known consumer uses NodeConverter directly,
so removing this class is not considered breaking.

All in all, this patch is considered internal and
*not* breaking as such.
